refactor(network): move identify event processing to own module
