+++ ant-networking/src/event/mod.rs
@@ -6,6 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
+mod identify;
 mod kad;
 mod request_response;
 mod swarm;
